não pode mais deletar o próprio login

+++ admin/users/delete.php
@@ -1,13 +1,17 @@
 <?php 
-require_once "../conexao.php";
-
+require_once "../../conexao.php";
+session_start();
 $id = isset($_GET['id']) ? (int) $_GET['id'] : null;
-$rota = $_GET['rota'] ?? null;
 
 if ($id) {
-    $mysqli->query("DELETE FROM admins WHERE admins_id = $id");
-    header("Location: ./");
-    exit;
+    if($id <> $_SESSION['adminId']){
+        $mysqli->query("DELETE FROM admins WHERE admins_id = $id");
+        header("Location: ./?msg=deleted");
+        exit;
+    }else{
+        header("Location: ./?msg=cannot_delete_self");
+        exit("Você não pode excluir sua conta.");
+    }
 } else {
     exit("ID inválido para exclusão.");
 }
